Support for multiple hosts to pbench-register-tool
Now the `--remote` parameter has been renamed to `--remotes` to denote
that it now accepts one or more remotes on which to register tools,
keeping the old `-r|--remote` syntax for compatibility.

Since labels are specified per host, we also add support for a new
`--labels` argument, checking to ensure that the number of labels
matches the number of remotes specified.

The user can also create a file containing a list of `<hostname>,<label>`
pairs (leading `#` comments and blank lines allowed) instead of listing
them all on the command line by just specifying `--remotes=@<filename>`.

We also add a new `--no-install` parameter for the user to opt out of
running the tool install, in support of environments where tools are
already pre-installed (e.g. container images).

The interface changes to add `--remotes`, `--labels`, and `--no-install`
applies to `pbench-register-tool-set` as well.

We enhance the unit tests to verify the new behaviors of the interfaces,
and also add a test for the `--group` parameter to
`pbench-register-tool-set`.

Fixes issue #1285.