Fixes for indexing
Reworked the handling of tools so that we have one method for handling
it and we avoid throwing stack traces as a result of missing tool data.

Change a few "error" log entries emitted by the indexer to "warnings"
since they don't represent an error with how the indexing is working.

Fixed a missing counters attribute of a `PbenchTemplate` and added a
few more uses of those counters.

Provided better names for all the "operational context" counters so
that we know where in the indexing of the data the counters were
created, while only printing out counters that have data.