Add support for indexing trafficgen results
We can now index result.json files generated by trafficgen.  We have
added a sample trafficgen tar ball which contains minimal tool data,
and a fairly reasonable amount of result data.  We have updated the
method of indexing to relax the names of the result arrays in the JSON
so that we can accommodate all the various data arrays from trafficgen.

This means that what we used to call "`measurement_units`" is now called
"`measurement_title`".  We also add a "`measurement_idx`" field to be
sure we track how many arrays we are processing, even though it is
usually only one.  Further, have fixed up the `mock` module to handle
more kinds of data types seen in the data sets we index.