Make s3 backup optional.
If the S3 service is inaccessible, the backup does not succeed and
keeps retrying. That may be reasonable if the S3 service is up and
running, but there may be situations where we want to bypass it.

Commenting out the [pbench-server-backup] section (or just the
endpoint_url option in that section) in the server config file now
allows the rest of the backup to succeed (including moving the
TO-BACKUP link to the final BACKED-UP state). That allows us to do
local backups, irrespective of the S3 backup service.

test-6.13 implements the commenting out of the endpoint_url option;
test-6.14 implements the commenting out of the [pbench-server-backup]
section (the tar ball is a symlink to the 6.13 tarball).