Enhance ansible dashboard README
We are updating the `README.md` for the dashboard to fix up a few
incorrect references to variable names expected by the `config.json.j2`
template.  We also adjust the sample `inventory` file.

Lastly, we ignore the `.yarnrc` and `.npmrc` files to avoid folks
inadvertently checking them into the code base.