Fix all bash CLIs to catch missing option args
The `getopt` does not detect when an option that expects a parameter
is missing when followed by another valid option.  For example, if
we have `getopt -o a:b:c`, then a command line of `prog -a -b -c` will
give `-b` as the option value for `-a`, and -b will not be considered
at all.

This has consequences for all bash based CLI processing that uses
`getopt` in the various pbench-agent modules.

Rather than fixing each one individually, we have written a drop-in
replacement for `getopt`, called `pbench-argparse`, which is a simple
Python 3 program to mimic the behavior of `getopt` while adding the
ability to check for options as arguments.