Remove redundant fetch of MD5s from verify backup
To remove the redundant fetch of the MD5 values we added those methods to
the BackupObject class which let's us store the accummulated values in the
object for use later.

We also enhance the report formating so that we get separartion between
sub-sections of the report, and a summary of the timings for all the
phases.