Audit now ignores incoming `*.unpack` dirs
The audit server code now ignores incoming `*.unpack` dirs which have valid
tar balls in the `$ARCHIVE` hierarchy, and reports those directories which
do not.

During review of the initial changes, we found a number of inconsistencies
with how temporary files are handled. We believe all of these have been
cleaned up with this commit.