Fix failing unit tests on some Fedora systems
It looks like newer versions of pass process associative arrays in
different orders.  So we consistently use a numeric indexing scheme to
avoid that behavior, allowing the unit tests to pass in all cases.

We fix a spelling error in a variable name the resulted in the correct
operation for the handling of the `--no-install-` option (thanks Nick).

We also add the line number where a remotes file is not formatted as
expected, and fix the debug log file output of `tool_opts` to use a
quoted string to be explicit.