Fix use of scp/ssh options and mock scp cleanly
The scp mock command now only mocks local copies, which are used when copying
remote tool data locally.  The other copies are only recorded.

There was a wayward tmp directory in the sample data which should not be there.

Finally, we fix the various scp commands to use the right options.