Refactor code base on operational review
First we change all uses of `[ ]` to `[[ ]]` so that we are using Bash
Shell syntax everywhere instead of Bourne shell, and we use the `${}`
form of variable name references to avoid problems in context when
expanding.

Second we remove redundant code, and fix up the permissions before we
move the tree into place.  We also do a better job of using the variables
we have defined, and cleaning up when results and users links encounter
errors.