Break pbench unpack tar balls up into four sizes
We now have four different jobs by default for processing tar ball sizes. Based
on historical pbench result sizes, we set default buckets to capture the first
90%, then next 5% (95th percentile), then the next 4% (99th percentile), and then
finally the last 1% (100th precentile).

We move the list generation into a function, and the main loop into a function in
preparation for further changes to make it easier to review.