Implement restart of unpack tar balls
The algorithm of unpack tar balls now stops processing the list of tar
balls to unpack if the time taken to unpack tar balls so far is twice as
long as it takes to find tar balls to unpack.  If the time taken to find
tar balls takes less than one minute, we just use 60 sec as the minimum
since that is the shortest possible interval of time between cronjob
invocations.

Further tar balls are processed now from newest to oldest, so by
restarting periodically, more recent results will get unpacked before
older results within a size bucket.

Removed the redundant test-5.1.setup file replacing it with a symlink
to test-5.setup (updating test-5.2.setup to link to the same).

Fixes issue #1398.