Log messages using logging functions in bash scripts.
Fixes #1421

Server bash scripts used "echo" and output redirection to log info and error messages to local files.
This PR does two things:

- It defines two functions, "log_info" and "log_error", in pbench-base.sh.
- It replaces uses of "echo" for logging purposes with one or the other of these two functions.

Note that the messages have not been changed at all and neither have the destinations: after the
change, the same messages are logged to the same places. This means that there is *no* change to
the gold files used for unit testing.

In a future PR, we are planning to change the logging functions in pbench-base.sh, so that they
follow the policy that is already used by server python scripts: logging to a centralized log
service. That is going to involve changes to pbench-base.sh and virtually all the gold files,
but it will involve *no* changes to any of the bash scripts.