Properly set the benchmark_run_dir for all tests
We need to ensure that the environment all the util-scripts are tested
under closely matches the environment in which they are actually invoked
(via bench-scripts, etc.). We were being pretty loose with the directory
used, so this commit tightens that down a bit.

We also make sure we run all unit tests by changing the current working
directory to the test environment we setup.

We consistently use `[[` conditional form, and we also use `${}` variable
references everywhere.