Add Marker to sosreport if Hostname not found
Fixes #1263

Earlier the process of indexing used to break then and there if
the sosreport of a tarball does not consist of a hostname. This
PR fixes this issue by adding an error message in the sosreport
regarding the issue with the hostname and completing the processing
of the tarball.