Refactor tool-scripts to use one script
Addresses issue #1459 following "Proposal 1".

This is a fairly invasive and major change to the tool sub-system.
The interface to the `pbench-start/stop/postprocess-tools` has
changed, and changes to various interfaces and behaviors of the
individual tools.

The first change is that the pbench-agent no longer tries to install
*ANY* RPMs via yum.

This is significant in that all of the benchmark scripts now only
*check* that their required RPMs are installed, and all the tools
during registration only perform a check that their requirements are
satistified.

Again, no RPMs are installed by the pbench-agent any longer.

All of the tool datalogs have been enhanced to check for their
required environment and error messages describing what is not met.

We are removing knowledge of how bench-scripts handle iterations
and samples so that the tools are only given a directory to work
with and no longer use `$pbench_tmp`.  All that they need to do
with files are managed in that given directory. This then allowed
us to remove the `--iteration` and `--group` parameters from all
the tools, and in turn, remove the --iteration parameter from the
invocation of the `pbench-start/stop/postprocess-tools`.  This
prevents knowledge of how benchmark scripts work from leaking into
the various tools.

Further, we now have three scripts that implement the methods for
running tools:

 * `base-tool`
    * a merger of all the previous tool scripts
    * with #1248, this will be replaced by Tool Meister code
 * `user-tool`
    * dedicated to just handling the user
 * `external-data-source`
    * place-holder for the no-op external data source

We also add the notion of a "stop-postprocess" script so that when
tools are stopped (when `pbench-stop-tools` is invoked), any
post-processing required to be performed on the host on which the
tool is running can be executed.  This allows us to run all the
normal post-procesing of data to generate charts and graphs during
the `pbench-postprocess-tools` phase only on the local controller, or
at a later time.

We have added unit tests for all the tools in the `tool-scripts`
directory, and in the `tool-scripts/datalog` directory as well.

One major behavior change to the `systemtap` tool is worth noting.
The systemtap script argument is now the "compiled" systemtap
script. The user is not responsible for making sure the provided
script is available to the systemtap tool on all the hosts it is
registered.

This is a work-in-progress!

We still have a few tools to re-factor to support:

 * [ ] haproxy-ocp
 * [ ] oc
 * [ ] pcp
 * [ ] pprof
 * [ ] prometheus-metrics

Unit test considerations:

 * [ ] consider how to unit-test the new *-stop-postprocess scripts