Add --iteration-list option to pbench-user-benchmark
This endows `pbench-user-benchmark` with the ability to accept a file
containing the list of iterations to run given the benchmark on the
command line.  The file contains an iteration name separated by a space
from the rest of the line, which is treated as the arguments to add to
the benchmark command invocation.

We have also modified `pbench-specjbb2005` with the same understanding
of the default iteration names used by `pbench-tool-trigger`.