Delay formatting logging strings.
For efficiency, we let the logging system do the formatting (or not).
That way, if the string is never logged, the formatting will not happen.

This commit fixes the logging calls in bin/pbench-backup-tarballs.py
and lib/s3backup/__init__.py. This is an opprtunistic change, since we
are going to fix bugs in these two files in a subsequent commit of
this PR.