Add methods to read pbench configuration files
PbenchConfig contains methods to help manipulate the pbench
configuration files. It uses configtools to load the configuration
files and configparser to read the configuration file.

This was done so that pbench agent commands and the pbench server
can use the same methods can to read, query, and display
configuration files.

Signed-off-by: Charles Short <chucks@redhat.com>