S3 backup: Large object fixes
Various bug fixes in backup, verify and the library.
A substantial amount of debug logging has been added to
the code.

The mock implementation of S3 was modified substantially to fix
a bunch of problems. Basically, it was not behaving like the real
thing and the differences were causing problems. The details can
be found in the README file which will be in the last commit of
this series.

The most serious problem was that the mock get_object() was wrongly
assuming that the object exists already. It is now modified to raise a
NoSuchKey exception. The generic get_tarball_header() routine is
modified to do impedance matching for the exceptions: the real S3
client raises ClientError and then we have to look in the response to
see what kind of error it is, but if it is a NoSuchKey error, we raise
our own NoSuchKey exception in that case as well.  backup_to_s3() then
catches that exception and acts accordingly.

Mock S3 implementation README file
----------------------------------
The mock implementation is described in some detail. Examples of the
real S3 implementation, together with some descriptions of its quirks,
are provided as motivation for the mock implementation.

The file should be helpful in understanding the unit tests (old and
new).

Unit test changes
-----------------

Gold files required changes because of additional debug logging in
the backup scripts.

The test-6.* gold files required additional changes because of fixes
to the mock S3 connector code in the library (see the s3backup/README
file for details).

The test-9.* gold files (verify backup) required additional changes
caused by bugs in the verify script, as well as the fixes to the mock
S3 connector code in the library (see the s3backup/README file for
details).

New tests for large object functionality
----------------------------------------

6.15 is derived from 6.12: the backups are already in place, so we
check that redoing the backup is idempotent (as long as the checksums
match).

6.16 and 6.17 test error conditions from the S3 server.

9.8 adds a large object test for testing the verify program.