Don't emit messages directly to the tty
The `pbench-fio` benchmark script was emitting a message directly to the
tty when the `fio` version could not be determined.  Instead we add it
to the `pbench.log` and emit it on `stderr`.

We also define the `PYTHONPATH` once in the server unit tests, and make
the python markers in the `getconf.py` modules uniform with other files.