Enhance indexing to handle pbench-run-benchmark
Partial fix for #1512.

The current way the indexer recognizes what to index or not is via the
`script` option of the `[run]` section in a `metadata.log` in a pbench
result.  We currently only accept `fio`, `uperf`, and `trafficgen`, the
names that `pbench-(fio|uperf|trafficgen)` emit today as the
`$benchmark` environment variable value.

However, when `pbench-run-benchmark` executes, the value it uses for
the `$benchmark` enviroment variable is the existing `pbench-*`
counterparts.

We add a test for a `fio` benchmark run via `pbench-run-benchmark` to
verify that.

This particular data set also uses the "user" feature, which identified
that we were missing a mapping for that field in `result-data-sample`,
and `tool-data-*` mappings.

We add proper handling and mappings for the `raw_size` and
`tar-ball-creation-timestamp` fields in the `@metadata` field of a
`run` document.

We add new fio fields to the mappings we track during unit tests.