Fix getconf.py and activation script invocations.
getconf.py is now part of the server bin/ directory and the
configtools library is in the lib/ directory, so we need to set PATH
(to pick up getconf.py) and PYTHONPATH (to pick up configtools)
appropriately.

In RHEL7, we need to use scl to invoke getconf.py and the activation
script that creates the crontab.