Add flake8 support
Add flake8 configuration file so that we can start
to correct pep8 errors.

In this commit to start we ignore everything because we are
only concerned with the infrastructure that is required
to implement lint checking.

For the developer we add a pre-commit hook so that after
each change flake8 will run on the code that has changed.
We also added a travis job to catch any linting errors
if the user did not have pre-commit-hooks installed locally.

Added a contributor guide on how to get started.

Signed-off-by: Charles Short <chucks@redhat.com>