Create a virtualenv for unittests
Before moving the python bits for pbench to a unified namespace
we need to update the unittests that pbench runs so that it runs
in a clean environment.

To do this we use python's virtualenv to set up the python environment
so that so dont conflict with python libraries installed on the host.
This means that if pbench-server/pbench-agent is intalled on the host
then the unittests wont fail when the pbench python library is installed
on the host.

This is a no-op right now because the unittests set their own $PYTHONPATH,
however this will be removed in a future commit that will have all of
the python bits in one namespace.

Signed-off-by: Charles Short <chucks@redhat.com>