Add support for linpack workloads
The changes here are not specifically for linpack, except for the updates
we made to the mappings.

We did find that the "mean" field needs to always be a double, and that
there are cases where the timeseries provided values don't have a
timestamp on them.  In that case, we use the timestamp of the start of
the entire pbench run as the timestamp for the value.  We also use check
and handle a missing timeseries field entirely.