Refactor indexer module functions into class methods
This is a code refactoring of the indexer module to reduce the number of
module level functions that effectively operation as class methods. That
is, all the functions took a `PbenchTarBall` object via `ptb` variables.

As a result of this clean up, further cleanup within classes can be
performed without cascading effects to all the function/method interfaces.

No functionality was changed.