Fix tool stop/postprocess handling
We first verify `stop` and `postprocess` tool actions have a tool
directory that exist.

The `stop` and `postprocess` actions are now captured in their own log
files in the `${tool_output_dir}`.  That reduces unnecessary output now
that `tool-scripts` operations are not captured in the `pbench.log` file
as of 53dba5f (PR #1460).

We also remove unneeded `tmp` directory trees in the samples.