Fix pbench-specjbb2005's iteration handling
The `pbench-metadata-log` `end` operation looks for a `.iterations`
file.  Without that file, we won't index the iteration names properly.
We now use `.iterations` liek all the rest of the benchmarks.

We also record the iteration name instead of the number in the list
of iterations, while adding the iteration number to the recorded data.