Add setuptools support
Add a setup.py to allow the user/developer to install pbench
as a python package. This commits allows pbench to take advtange
of distutils, which is the standard of distributing Python modules.

This will utlimately will make pbench easier to package as an RPM
and allow it to be installed as a python package as well.

Signed-off-by: Charles Short <chucks@redhat.com>