Update unittests
Fix traceback while running the unittests with python-click:

Traceback (most recent call last):
  ...
RuntimeError: Click will abort further execution because Python 3 was
  configured to use ASCII as encoding for the environment. Either switch
  to Python 2 or consult http://click.pocoo.org/python3/ for
  mitigation steps.

This is due to the fact that the unit tests are running with python3
with restricted to ASCII data. The solution to these problems is
different depending on which locale your computer is running.

A simple workaround for this is to set LC_ALL and LANG to C.UTF-8
while running the unit tests.

Signed-off-by: Charles Short <chucks@redhat.com>