Added script specific and default logging level
Fixes issue #1530

With this there can be a logging level that can be specific to any
script, we just need to mention that in the config file. If nothing
is mentioned than there is always a default logging level that can
be used.

Test for checking the working of logging-level is also added in this
commit.

(cherry picked from commit 8572c9793e03e75a18332bfa78ec74e436ef9ca0)