Run unittests with tox
Tox aims to automate and standardize testing in Python. It is part of a
larger vision of easing the packaging, testing, and release process of
pbench.

Here, we setup tox environments for lint, py3, server, and agent to
allow users and developers unit tests and linting. We automate the
installation of required python dependencies through requirements.txt
and test-requirements.txt

Updated the .travis.yml to use tox.

Also updated contributors guide to include how to install tox, and to
run the unittests.

Signed-off-by: Charles Short <chucks@redhat.com>