Make getconf.py installable by using setuptools
Unify the getconf.py for both server and agent so that we are using only
one script to read the configuraiton file. This is done by detecting
if the PBENCH_AGNET_CONFIG is set or if the PBENCH_SERVER_CONFIG is
set via envrionment variables.

Next make the getconf.py installable via setuptools. This was done
by adding the "pbench-agent" binary to the setup.cfg. "pbench-config"
was chosen for the name of the config because getconf was already being
used by glibc and breaking people's systems is never a good idea.

Finally add apprioate unit tests for the new python code.

Signed-off-by: Charles Short <chucks@redhat.com>