Move to support uperf-1.0.7
We rename the RPM we are expecting for `uperf` from `pbench-uperf to
just `uperf`.

We move pbench-uperf to fetch the version information of the `uperf`
RPM from the `pbench-agent.cfg` file like we do for `pbench-fio`.

We also fix up the test infrastructure to use the
`pbench-agent-default.cfg` following the DRY principle.