Use mv instead of cp to avoid a data copy
If the reception area and the archive area are on the same volume,
using a `cp` command will ALWAYS copy the data. But if we use a `mv`
command then we can avoid a data copy when the reception and archive
areas share a volume.  When the reception and archive areas are on
separate volumes, a `mv` will just do a "copy-n-delete".

We change the algorithm to first copy the `.md5` file to the
destination so that we have a well defined set of clean steps when
the tar ball copy fails.