Address indexing related behaviors seen in the wild
First, we have seen the `action_retry_deque` assertion fire, which was
fairly useless as it was because no context information was given. So
we address that by turning that assertion into a logging "error",
adding some additional context information to help.

Second, we add the tar ball being indexed to error and exception log
messages to make it more convenient to track down the problem when
debugging.

Third, there is a problem deriving host names given IP addresses, so
we fixed the code to make sure the logic is correct and works in all
cases. We also added a new test, test-7.25, to verify the behavior.

Finally we fix a small accounting bug where we were not incrementing
the counter for the number of result data documents yielded!