Add pbench-config cli
Unify the server/agent getconf.py so that we arent re-using the same
code with minimal differene. Also add unittests to test output.

Aso update the unit testing framework to add a "functional"
envrionment that we can check to see if the scripts that are
installed via setuptools are working correctly.

Signed-off-by: Charles Short <chucks@redhat.com>