Add server maint role, drop tool indexing
We add the notion of a maintenance role for the audit service, and
drop the tool data indexing by default.

We add commented out `pbench-server-backup` and `Indexing` sections
to the default server configuration file with a comment that they
are deliberately omitted from the defaults.  This avoids making any
code changes to accommodate always having a default.

We drop the default "vos" user for sosreports and use the default user.
This let's us remove a few unnecessary config files and their symlinks
from the unit tests.

We also take this opportunity to make sure all the configuration
files in the tree a properly sync'd with each other.