Fix server ansible roles.
pbench-clean-yum-cache role: (new)
Add a pbench-clean-yum-cache role: the ansible package module is not
always successful in updating to the latest package, particularly on RHEL7
boxes. This is an attempt to help it do that, but it remains to be seen how
effective it is.

pbench-server-install role:
Import the pbench-clean-yum-cache role so that it is executed before we try
to install/update packages.

pbench-server-user role:
Add some  SELinux settings to the /home/<pbench>/.ssh directory.

Fix comments and add pbench_repo_url_prefix variable to the inventory example
file.