Fixes from review plus additional ones
pbench-clean-yum-cache role:
Delete /var/cache/yum directory.

pbench-server-install-config-file role:
There is a chicken-and-egg problem here: before we install the config file
we don't know what pbench_owner is going to be - nevertheless we installed it
as owned by "pbench.pbench", which is OK if the user/group exists, but if they
don't, that gives an error. Install it as owned by "root.root" and fix it up
later.

pbench-server-user role:
Change the ownership of the config file(s).

pbench-server-install role:
Fix import_role syntax error.

pbench-server-change-ownership role:
New role to change the ownership of the installation tree to whatever user/group
is specified in the config file. The RPM installs everything as root.

pbench-server-config role:
Invoke the pbench-server-change-ownership role above as the last thing we do
for configuration.