Refactoring to introduce PbenchAgentConfig class
Leverages more shared code, we have created a `PbenchAgentConfig` class
which is a sub-class of the common `PbenchConfig` class, allowing us to
share more code between the server and agent environments.

Having the `PbenchConfig` class shared with the agent side as well allows
us to introduce the logging infrastructure, previously only available to
the server side, to the agent side as well.

Along the way, we address a number of issues in the Python3 code found
while re-working unit tests, including the proper use of temporary
directories.