RPM making for pbench-agent
Fix the agent Makefile to install the tools, but not tests, gold files
etc.  We also now generate `pbench-config` for the agent side from the
`python3 setup.py install` operation.

Get rid of most of the cruft of the old version: move a couple of things
from the old misc-scripts into a new utils/ sudbir and get rid of
`misc-scripts` altogether.

Get rid of the `MANIFEST` file since we don't use it any longer.

Add some missing stuff to the agent Makefile install target (in
particular, the bench-scripts/postprocess scripts).

We take a page from the `b0.69` branch, and refactor the Makefile to avoid
having to build the source tree twice, once for the source tar ball, once
when running the RPM build.  Now we build the install tree for the source
tar ball and just use that directly.

We also refactor the agent source tree `Makefile` so that we only add what
we need to the source tar ball, allowing us to drop the use of the
`exclude.pats` file.

Further, we switch to using a Jinja2 template for the `pbench-agent.spec`
file, allowing us to drop the need for the `set-version-release` script.