RPM making for pbench-agent
[Back-port with adjustments]

Fix the agent Makefile to install the tools, but not tests, gold files
etc.

Get rid of most of the cruft of the old version: move a couple of
things from the old misc-scripts into a new utils/ sudbir and get
rid of misc-scripts altogether.

Get rid of MANIFEST file: we don't use it any longer.

The python lib for the agent is outside the agent/ subdir to begin
with, but it is moved inside it on installation, so we have to have
special treatment for it.

Add some missing stuff to the agent Makefile install target
(in particular, the bench-scripts/postprocess scripts).