RPM making for pbench-agent
Fix the agent Makefile to install the tools, but not tests, gold files
etc.

Get rid of most of the cruft of the old version: move a couple of
things from the old misc-scripts into a new utils/ sudbir and get
rid of misc-scripts altogether.

Get rid of MANIFEST file: we don't use it any longer.

A few ugly hacks remain:

- the python lib for the agent is outside the agent/ subdir to begin
  with, but it is moved inside it on installation, so we have to have
  special treatment for it.
- the agent Makefile is installed to begin with because we need it
  in the install phase of the RPM, but we delete it afterwards since
  we don't need it in the final installation.

Add an explicit pbench-config.
Add some missing stuff to the agent Makefile install target
(in particular, the bench-scripts/postprocess scripts).