Avoid hardcoded tools location on bench-scripts
In case the benchmark_bin is not set the "pbench-fio" uses
"/usr/local/bin/fio" and "pbench-uperf" uses "/usr/bin/uperf". This is
not systematic and could lead to confusion, let's use which command to
obtain the right tool location.

To simplify potential future changes let's add a functions to get
benchmark_bin and version.

Co-authored-by: Lukáš Doktor <ldoktor@redhat.com>
Co-authored-by: Peter Portante <peter.portante@redhat.com>
Signed-off-by: Peter Portante <peter.portante@redhat.com>
Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>