Change ToC `parent` field to "keyword"
Queries made to find all JSON documents which have a `parent` field
with a specific value require the use of a "keyword" mapping
(`not_analyzed` field in Elasticsearch V1 terms).  When we had the
`path_analyzed` mapping in place, the query string was tokenized in
the same way such that some of the tokens would match, resulting in
more documents being returned in the query that did not have exactly
those tokens.

This resolves a problem for the `pbench-dashboard` code allowing the
Table-of-Contents functionality to properly query for documents whose
`parent` field matches exactly.

The gold files were updated using a simple `sed` command to record
the fact that the mapping version changed (only `test-7.0.1` was
updated by hand):

```
sed -i 's/pbench-unittests\.v4\.run/pbench-unittests.v5.run/' gold/test-*.txt
```

Note that because in Elasticsearch V1 the ToC JSON documents live
in the same index as the "run" documents, we have to update the run
document's version to match.