Prometheus/Node_Exporter Full V1 Integration Commit
This work adds the notion of a "collector" to the Tool Data Sink, and
"persistent tools" which run continuously without cycling through the
"start", "stop", and "send" phases.  The collector is responsible for
continuously pulling data from those tools which are now started during
the new "init" phase, and stopped during the new "end" phase.

The first actual implementation of this kind of collector is for the
Prometheus data collection environment, where a `node-exporter`
"persistent tool" is run providing an end-point for a prometheus server
"collector" to pull data from it and store it locally off the run
directory (`${benchmark_run_dir}`,
e.g. `${benchmark_run_dir}/collector/prometheus`).

This work adds the `tool-scripts/meta.json` file which is used to
describe which tools are persistent and which are transient (default).